Styled Content. Fades, Fonts, Colors
Customized the look of the overall site to make it looke cleaner and more comfortable for the eyes. The colors are now softer. The transitions are now all fades to allow for soft transitions taht are more pleasing to the eyes. The spacing is also mostly fixed. The "Correct" or "Try Again" font is yet to be colored though.